+++ src/components/AlbumsList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Album } from "../modules/posts/types";
-import {Card, Col, Row} from "antd";
+import { Album } from '../modules/albums/types';
+import {Card, Col, Row} from 'antd';
 
 interface AlbumsListProps {
     albumsList: Album[];
@@ -10,7 +10,7 @@ export const AlbumsList: React.FC<AlbumsListProps> = ({albumsList}) => (
         <Row>
             {albumsList.map(albums => (
                 <Col key={albums.id} span={6}>
-                    <Card title={albums.title} size='small'>
+                    <Card title={albums.title} size="small">
                         <span>{albums.userId}</span>
                     </Card>
                 </Col>
