+++ src/modules/albums/types.ts
@@ -1,0 +1,5 @@
+export interface Album {
+    userId: number;
+    id: number;
+    title: string
+}
